+++ test/ttmlir/Dialect/TTNN/permute/permute_tests_negative.mlir
@@ -6,7 +6,7 @@
 module {
   func.func @permute_non_valid_permutation(%arg0: tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16> {
     // CHECK: error: 'ttnn.permute' op Expected a permutation of {k | 0 <= k < 3} got (0, 1, 0)
-    %0 = "ttnn.permute"(%arg0) <{permutation = array<i32: 0, 1, 0>}> : (tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
+    %0 = "ttnn.permute"(%arg0) <{permutation = array<i64: 0, 1, 0>}> : (tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
     return %0 : tensor<16x32x64xbf16>
   }
 }
@@ -16,7 +16,7 @@ module {
 module {
   func.func @permute_non_valid_permutation(%arg0: tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16> {
     // CHECK: error: 'ttnn.permute' op Expected a permutation of {k | 0 <= k < 3} got (0, 1)
-    %0 = "ttnn.permute"(%arg0) <{permutation = array<i32: 0, 1>}> : (tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
+    %0 = "ttnn.permute"(%arg0) <{permutation = array<i64: 0, 1>}> : (tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
     return %0 : tensor<16x32x64xbf16>
   }
 }
@@ -27,7 +27,7 @@ module {
 module {
   func.func @permute_non_valid_permutation(%arg0: tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16> {
     // CHECK: error: 'ttnn.permute' op Expected result shape (16, 64, 32), got (16, 32, 64)
-    %0 = "ttnn.permute"(%arg0) <{permutation = array<i32: 0, 2, 1>}> : (tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
+    %0 = "ttnn.permute"(%arg0) <{permutation = array<i64: 0, 2, 1>}> : (tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
     return %0 : tensor<16x32x64xbf16>
   }
 }
